third_party/WebKit/LayoutTests/http/tests/devtools/console/console-log-side-effects-expected.txt

CONSOLE MESSAGE: line 16: string
CONSOLE MESSAGE: line 17: 42
CONSOLE MESSAGE: line 18: false
CONSOLE MESSAGE: line 19: undefined
CONSOLE MESSAGE: line 20: null
CONSOLE MESSAGE: line 21: NaN
CONSOLE MESSAGE: line 22: -Infinity
CONSOLE MESSAGE: line 23: 0
CONSOLE MESSAGE: line 24: 42.0000
CONSOLE MESSAGE: line 25: -4.24200e-11
CONSOLE MESSAGE: line 26: true
CONSOLE MESSAGE: line 27: foo
CONSOLE MESSAGE: line 28: [object Object]
CONSOLE MESSAGE: line 29: [object Window]
CONSOLE MESSAGE: line 35: [object HTMLParagraphElement]
CONSOLE MESSAGE: line 39: [object Object]
CONSOLE MESSAGE: line 40: 1,2,3
CONSOLE MESSAGE: line 42: [object Uint32Array]
CONSOLE MESSAGE: line 45: 1,2,3,[object Object]
CONSOLE MESSAGE: line 48: [object Object]
CONSOLE MESSAGE: line 51: 1,2,3,[object Object],[object Object]
CONSOLE MESSAGE: line 62: 1,2,3,4,5,,,,,,NaN,,,,,,,NaN,1,2,3,4,5,,
CONSOLE MESSAGE: line 71: [object Object]
CONSOLE MESSAGE: line 74: 1,2,3,[object Object],[object Object],[object Object]
CONSOLE MESSAGE: line 80: 
CONSOLE MESSAGE: line 80: 
CONSOLE MESSAGE: line 80: 
CONSOLE MESSAGE: line 80: 
CONSOLE MESSAGE: line 80: 
CONSOLE MESSAGE: line 83: 
CONSOLE MESSAGE: line 90: 
CONSOLE MESSAGE: line 96: 
Tests various extreme usages of console.log()
